// legacy/20240613-components-question-PasteForm/types.ts
export interface ITag {
  id: string;
  name: string;
  color: string;
}